+++ src/pages/auth/index.ts
@@ -1,3 +1,2 @@
 export * from "./ui";
-export * as authModel from "./model/model";
-export * as selectors from "./model/selectors";
+export * as authModel from "./model";
