--- src/pages/auth/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./model";
-export * as selectors from "./selectors";
