--- src/pages/auth/model/selectors.ts
@@ -1,5 +1,0 @@
-import { useUnit } from "effector-react";
-import { $emailField, $isValid, actions } from "./model";
-
-export const useEmailField = () =>
-  useUnit([$emailField, actions.changedEmail, $isValid]);
