--- src/pages/workspace/boards/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./model";
-export * as selectors from "./selectors";
