--- src/pages/workspace/boards/model/selectors.ts
@@ -1,9 +1,0 @@
-import { useUnit } from "effector-react";
-import { $boardsEmpty, $boardsLength, $isNotFound, $search, actions } from ".";
-
-export const useSearch = () =>
-  useUnit([$search, actions.searched, actions.resetSearch]);
-
-export const useBoardLength = () => useUnit([$boardsLength, $boardsEmpty]);
-
-export const useEmptySearchResult = () => useUnit($isNotFound);
